requirements.txt
ipykernel
pandas
numpy
matplotlib
seaborn
scipy
scikit-learn
imblearn
pymongo
certifi
python-dotenv
dill
PyYAML
boto3
botocore
fastapi
uvicorn
jinja2
python-multipart

# -e .
